Split FarePolicy domain types
