+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/dynamic-offer-driver-app.cabal
@@ -157,6 +157,11 @@ library
       Domain.Types.Exophone
       Domain.Types.FareParameters
       Domain.Types.FarePolicy
+      Domain.Types.FarePolicy.Common
+      Domain.Types.FarePolicy.DriverExtraFeeBounds
+      Domain.Types.FarePolicy.FarePolicyProgressiveDetails
+      Domain.Types.FarePolicy.FarePolicySlabsDetails
+      Domain.Types.FarePolicy.FarePolicySlabsDetails.FarePolicySlabsDetailsSlab
       Domain.Types.Geometry
       Domain.Types.Issue.Comment
       Domain.Types.Issue.IssueCategory
