+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Types/FarePolicy.hs
@@ -12,11 +12,12 @@
  the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
 -}
 
-module Domain.Types.FarePolicy where
+module Domain.Types.FarePolicy (module Reexport, module Domain.Types.FarePolicy) where
 
-import qualified Data.List.NonEmpty as NE
-import Data.Ord
 import Domain.Types.Common
+import Domain.Types.FarePolicy.DriverExtraFeeBounds as Reexport
+import Domain.Types.FarePolicy.FarePolicyProgressiveDetails as Reexport
+import Domain.Types.FarePolicy.FarePolicySlabsDetails as Reexport
 import qualified Domain.Types.Merchant as DM
 import qualified Domain.Types.Vehicle.Variant as Variant
 import Kernel.Prelude
@@ -53,66 +54,6 @@ instance FromJSON (FarePolicyDetailsD 'Unsafe)
 
 instance ToJSON (FarePolicyDetailsD 'Unsafe)
 
-data FPProgressiveDetailsD (s :: UsageSafety) = FPProgressiveDetails
-  { baseFare :: Money,
-    baseDistance :: Meters,
-    perExtraKmFare :: HighPrecMoney,
-    deadKmFare :: Money,
-    waitingCharge :: Maybe WaitingCharge,
-    nightShiftCharge :: Maybe NightShiftCharge
-  }
-  deriving (Generic, Show, Eq)
-
-type FPProgressiveDetails = FPProgressiveDetailsD 'Safe
-
-instance FromJSON (FPProgressiveDetailsD 'Unsafe)
-
-instance ToJSON (FPProgressiveDetailsD 'Unsafe)
-
-newtype FPSlabsDetailsD (s :: UsageSafety) = FPSlabsDetails
-  { slabs :: NonEmpty (FPSlabsDetailsSlabD s)
-  }
-  deriving (Generic, Show, Eq)
-
-type FPSlabsDetails = FPSlabsDetailsD 'Safe
-
-instance FromJSON (FPSlabsDetailsD 'Unsafe)
-
-instance ToJSON (FPSlabsDetailsD 'Unsafe)
-
-findFPSlabsDetailsSlabByDistance :: Meters -> NonEmpty (FPSlabsDetailsSlabD s) -> FPSlabsDetailsSlabD s
-findFPSlabsDetailsSlabByDistance dist slabList = do
-  case NE.filter (\slab -> slab.startDistance < dist) $ NE.sortBy (comparing (.startDistance)) slabList of
-    [] -> error $ "Slab for dist = " <> show dist <> " not found. Non-emptiness supposed to be guaranteed by app logic."
-    a -> last a
-
-data FPSlabsDetailsSlabD (s :: UsageSafety) = FPSlabsDetailsSlab
-  { startDistance :: Meters,
-    baseFare :: Money,
-    waitingCharge :: Maybe WaitingCharge,
-    nightShiftCharge :: Maybe NightShiftCharge
-  }
-  deriving (Generic, Show, Eq, ToSchema)
-
-type FPSlabsDetailsSlab = FPSlabsDetailsSlabD 'Safe
-
-instance FromJSON (FPSlabsDetailsSlabD 'Unsafe)
-
-instance ToJSON (FPSlabsDetailsSlabD 'Unsafe)
-
-data DriverExtraFeeBounds = DriverExtraFeeBounds
-  { startDistance :: Meters,
-    minFee :: Money,
-    maxFee :: Money
-  }
-  deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
-
-findDriverExtraFeeBoundsByDistance :: Meters -> NonEmpty DriverExtraFeeBounds -> DriverExtraFeeBounds
-findDriverExtraFeeBoundsByDistance dist driverExtraFeeBoundsList = do
-  case NE.filter (\driverExtraFeeBounds -> driverExtraFeeBounds.startDistance < dist) $ NE.sortBy (comparing (.startDistance)) driverExtraFeeBoundsList of
-    [] -> error $ "DriverExtraFeeBounds for dist = " <> show dist <> " not found. Non-emptiness supposed to be guaranteed by app logic."
-    a -> last a
-
 data NightShiftBounds = NightShiftBounds
   { nightShiftStart :: TimeOfDay,
     nightShiftEnd :: TimeOfDay
@@ -123,12 +64,6 @@ data AllowedTripDistanceBounds = AllowedTripDistanceBounds
   { maxAllowedTripDistance :: Meters,
     minAllowedTripDistance :: Meters
   }
-  deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
-
-data WaitingCharge = PerMinuteWaitingCharge HighPrecMoney | ConstantWaitingCharge Money
-  deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
-
-data NightShiftCharge = ProgressiveNightShiftCharge Float | ConstantNightShiftCharge Money
   deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
 
 data FarePolicyType = Progressive | Slabs deriving (Show, Read)
@@ -157,29 +92,6 @@ data FarePolicyAPIEntity = FarePolicyAPIEntity
 data FarePolicyDetailsAPIEntity = ProgressiveDetailsAPIEntity FPProgressiveDetailsAPIEntity | SlabsDetailsAPIEntity FPSlabsDetailsAPIEntity
   deriving (Generic, Show, ToJSON, FromJSON, ToSchema)
 
-data FPProgressiveDetailsAPIEntity = FPProgressiveDetailsAPIEntity
-  { baseFare :: Money,
-    baseDistance :: Meters,
-    perExtraKmFare :: HighPrecMoney,
-    deadKmFare :: Money,
-    waitingCharge :: Maybe WaitingCharge,
-    nightShiftCharge :: Maybe NightShiftCharge
-  }
-  deriving (Generic, Show, ToJSON, FromJSON, ToSchema)
-
-newtype FPSlabsDetailsAPIEntity = FPSlabsDetailsAPIEntity
-  { slabs :: NonEmpty FPSlabsDetailsSlabAPIEntity
-  }
-  deriving (Generic, Show, ToJSON, FromJSON, ToSchema)
-
-data FPSlabsDetailsSlabAPIEntity = FPSlabsDetailsSlabAPIEntity
-  { startDistance :: Meters,
-    baseFare :: Money,
-    waitingCharge :: Maybe WaitingCharge,
-    nightShiftCharge :: Maybe NightShiftCharge
-  }
-  deriving (Generic, Show, Eq, FromJSON, ToJSON, ToSchema)
-
 makeFarePolicyAPIEntity :: FarePolicy -> FarePolicyAPIEntity
 makeFarePolicyAPIEntity FarePolicy {..} =
   FarePolicyAPIEntity
@@ -191,21 +103,3 @@ makeFarePolicyAPIEntity FarePolicy {..} =
     makeFarePolicyDetailsAPIEntity = \case
       ProgressiveDetails det -> ProgressiveDetailsAPIEntity $ makeFPProgressiveDetailsAPIEntity det
       SlabsDetails det -> SlabsDetailsAPIEntity $ makeFPSlabsDetailsAPIEntity det
-
-    makeFPProgressiveDetailsAPIEntity :: FPProgressiveDetails -> FPProgressiveDetailsAPIEntity
-    makeFPProgressiveDetailsAPIEntity FPProgressiveDetails {..} =
-      FPProgressiveDetailsAPIEntity
-        { ..
-        }
-
-    makeFPSlabsDetailsAPIEntity :: FPSlabsDetails -> FPSlabsDetailsAPIEntity
-    makeFPSlabsDetailsAPIEntity FPSlabsDetails {..} =
-      FPSlabsDetailsAPIEntity
-        { slabs = makeFPSlabsDetailsSlabAPIEntity <$> slabs
-        }
-
-    makeFPSlabsDetailsSlabAPIEntity :: FPSlabsDetailsSlab -> FPSlabsDetailsSlabAPIEntity
-    makeFPSlabsDetailsSlabAPIEntity FPSlabsDetailsSlab {..} =
-      FPSlabsDetailsSlabAPIEntity
-        { ..
-        }
