+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Types/FarePolicy/Common.hs
@@ -1,0 +1,24 @@
+{-
+ Copyright 2022-23, Juspay India Pvt Ltd
+
+ This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License
+
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version. This program
+
+ is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+
+ or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. You should have received a copy of
+
+ the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
+-}
+
+module Domain.Types.FarePolicy.Common where
+
+import Kernel.Prelude
+import Kernel.Types.Common
+
+data WaitingCharge = PerMinuteWaitingCharge HighPrecMoney | ConstantWaitingCharge Money
+  deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
+
+data NightShiftCharge = ProgressiveNightShiftCharge Float | ConstantNightShiftCharge Money
+  deriving (Generic, Eq, Show, ToJSON, FromJSON, ToSchema)
