+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Types/FarePolicy/FarePolicyProgressiveDetails.hs
@@ -1,0 +1,60 @@
+{-
+ Copyright 2022-23, Juspay India Pvt Ltd
+
+ This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License
+
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version. This program
+
+ is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+
+ or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details. You should have received a copy of
+
+ the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
+-}
+
+module Domain.Types.FarePolicy.FarePolicyProgressiveDetails
+  ( module Reexport,
+    module Domain.Types.FarePolicy.FarePolicyProgressiveDetails,
+  )
+where
+
+import Domain.Types.Common
+import Domain.Types.FarePolicy.Common as Reexport
+import Kernel.Prelude
+import Kernel.Types.Common
+
+data FPProgressiveDetailsD (s :: UsageSafety) = FPProgressiveDetails
+  { baseFare :: Money,
+    baseDistance :: Meters,
+    perExtraKmFare :: HighPrecMoney,
+    deadKmFare :: Money,
+    waitingCharge :: Maybe WaitingCharge,
+    nightShiftCharge :: Maybe NightShiftCharge
+  }
+  deriving (Generic, Show, Eq)
+
+type FPProgressiveDetails = FPProgressiveDetailsD 'Safe
+
+instance FromJSON (FPProgressiveDetailsD 'Unsafe)
+
+instance ToJSON (FPProgressiveDetailsD 'Unsafe)
+
+-----------------------------------------------------------------------------------------------------------------------------------------
+------------------------------------------------APIEntity--------------------------------------------------------------------------------
+-----------------------------------------------------------------------------------------------------------------------------------------
+
+data FPProgressiveDetailsAPIEntity = FPProgressiveDetailsAPIEntity
+  { baseFare :: Money,
+    baseDistance :: Meters,
+    perExtraKmFare :: HighPrecMoney,
+    deadKmFare :: Money,
+    waitingCharge :: Maybe WaitingCharge,
+    nightShiftCharge :: Maybe NightShiftCharge
+  }
+  deriving (Generic, Show, ToJSON, FromJSON, ToSchema)
+
+makeFPProgressiveDetailsAPIEntity :: FPProgressiveDetails -> FPProgressiveDetailsAPIEntity
+makeFPProgressiveDetailsAPIEntity FPProgressiveDetails {..} =
+  FPProgressiveDetailsAPIEntity
+    { ..
+    }
